// src/models/Car/Request/CreateCarRequest.ts
export interface CreateCarRequest {
    kilometer: number;
    year: number;
    dailyPrice: number;
    plate: string;
    minFindeksRate: number;
    imagePath: string;
    passengerCapacity: number;
    category: string;
    gearType: string;
    fuelType: string;
    colorId: number;
    modelId: number;
}